移植LFS

+++ fw/src/hal/hal_spi_bus.h
@@ -4,24 +4,21 @@
 #include <stddef.h>
 #include <stdint.h>
 
-
 typedef struct {
-    void* p_rx_buffer;
-    size_t rx_length;
-    void* p_tx_buffer;
-    size_t tx_length;
-}spi_transaction_t;
+    void *p_rx_buffer;
+    uint32_t rx_length;
+    void *p_tx_buffer;
+    uint32_t tx_length;
+} spi_transaction_t;
 
 typedef struct {
     uint32_t cs_pin;
 } spi_device_t;
 
-
 void hal_spi_bus_init();
-void hal_spi_bus_attach(spi_device_t* p_dev);
-void hal_spi_bus_aquire(spi_device_t* p_dev);
-void hal_spi_bus_release(spi_device_t* p_dev);
-uint32_t hal_spi_bus_xfer(spi_transaction_t* p_trans);
+void hal_spi_bus_attach(spi_device_t *p_dev);
+void hal_spi_bus_aquire(spi_device_t *p_dev);
+void hal_spi_bus_release(spi_device_t *p_dev);
+uint32_t hal_spi_bus_xfer(spi_transaction_t *p_trans);
 
-
-#endif 
+#endif
